agent-blueprint/serverless-mcp-farm/tavily-web-search/src/requirements.txt
tavily-python
awslabs-mcp-lambda-handler
requests
